Some bugs fixed @whatthecommit.com
